refactor: Extract route handlers to controller    classes
  - Create AuthController, UserController, and
  MovieController
  - Move all route handler logic to dedicated
  controller classes
  - Remove unused imports and clean up app.ts
  structure
